refactor: move useFeelingsData hook to hooks folder
